requirements.txt
langchain
langchain-community
llama-parse
fastembed
qdrant_client
python-dotenv
langchain-groq
chainlit
fastembed
langchain_cohere
sentence-transformers
unstructured[md]